tool/requirements.txt
cairosvg
Pillow